Added proper cast when determining key length for hash

+++ src/hash_table.c
@@ -87,7 +87,7 @@ void hash_table_delete_table(hash_table_table * hash_table) {
 // define hashing function
 static int hash_table_hash(const char * key, const int prime, const int hash_table_size) {
     long hash_value = 0;
-    const int key_length = strlen(key);
+    const int key_length = (const int) strlen(key);
     for (int i = 0; i < key_length; i++) {
         hash_value += (long) pow(prime, key_length - (i + 1)) * key[i];
         hash_value = hash_value % hash_table_size;
